remove unused page
