+++ mobile/lib/routing/router.dart
@@ -52,7 +52,6 @@ import 'package:immich_mobile/pages/search/map/map_location_picker.page.dart';
 import 'package:immich_mobile/pages/search/person_result.page.dart';
 import 'package:immich_mobile/pages/search/recently_added.page.dart';
 import 'package:immich_mobile/pages/search/search.page.dart';
-import 'package:immich_mobile/pages/search/search_input.page.dart';
 import 'package:immich_mobile/pages/library/partner/partner.page.dart';
 import 'package:immich_mobile/pages/library/partner/partner_detail.page.dart';
 import 'package:immich_mobile/pages/library/shared_link/shared_link.page.dart';
@@ -106,7 +105,7 @@ class AppRouter extends RootStackRouter {
           guards: [_authGuard, _duplicateGuard],
         ),
         AutoRoute(
-          page: SearchInputRoute.page,
+          page: SearchRoute.page,
           guards: [_authGuard, _duplicateGuard],
           maintainState: false,
         ),
@@ -243,11 +242,6 @@ class AppRouter extends RootStackRouter {
     AutoRoute(
       page: BackupOptionsRoute.page,
       guards: [_authGuard, _duplicateGuard],
-    ),
-    CustomRoute(
-      page: SearchInputRoute.page,
-      guards: [_authGuard, _duplicateGuard],
-      transitionsBuilder: TransitionsBuilders.noTransition,
     ),
     AutoRoute(
       page: HeaderSettingsRoute.page,
